services/ml/requirements.txt
kafka-python==2.0.2
scikit-learn==1.5.0
mlflow==2.13.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
